+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vsce-gitkeep" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
 
-- Initial release
+## [1.0.0] - 2019-12-03
+### Added
+- Insert .gitkeep file automatically as soon as a directory becomes empty.
+- Remove .gitkeep file automatically whenever a directory is populated with real files/directories.
+- Enabled automatically for git repository workspace (with a `.git` folder).
+- Can detect `.gitignore` file changes and ignore directories listed in it.
+- A simple command is provided to enable the extension for non-git-repository workspace.